+++ src/otx/algo/classification/multiclass_models/vit.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 import types
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Generic
+from typing import TYPE_CHECKING, Any, Callable
 from urllib.parse import urlparse
 
 import numpy as np
@@ -54,7 +54,7 @@ pretrained_urls = {
 }
 
 
-class ForwardExplainMixInForViT(Generic[T_OTXBatchPredEntity, T_OTXBatchDataEntity]):
+class ForwardExplainMixInForViT:
     """ViT model which can attach a XAI (Explainable AI) branch."""
 
     explain_mode: bool