+++ src/otx/algo/instance_segmentation/maskrcnn_tv.py
@@ -3,6 +3,8 @@
 #
 """TV MaskRCNN model implementations."""
 
+# type: ignore[override]
+
 from __future__ import annotations
 
 from typing import Any, ClassVar